backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
